Refactor code (#8)
* Move cli to its own module

* Fix test

* Move input and set_storage

* Move seal0_value_transferred

* Move seal_return

* Add new modules and types

* Create executor

* Delete unused traits
