+++ src/utils/types.rs
@@ -1,0 +1,2 @@
+pub type AccountId = [u8; 32];
+pub type Balance = u128;
