+++ src/wasm/mod.rs
@@ -1,0 +1,4 @@
+pub mod executor;
+pub mod host_functions;
+pub mod host_state;
+pub mod module;
